refactor(registry): re-work registry package arrangement
