+++ api/browse.go
@@ -8,20 +8,20 @@ import (
 	"github.com/qri-io/qri/lib"
 )
 
-// FeedHandlers provides HTTP API handlers for fetching content
+// BrowseHandlers provides HTTP API handlers for fetching content
 // from the qri network
-type FeedHandlers struct {
-	lib.FeedMethods
+type BrowseHandlers struct {
+	lib.BrowseMethods
 }
 
-// NewFeedHandlers allocates a FeedHandlers pointer
-func NewFeedHandlers(inst *lib.Instance) *FeedHandlers {
-	req := lib.NewFeedMethods(inst)
-	return &FeedHandlers{*req}
+// NewBrowseHandlers allocates a BrowseHandlers pointer
+func NewBrowseHandlers(inst *lib.Instance) *BrowseHandlers {
+	req := lib.NewBrowseMethods(inst)
+	return &BrowseHandlers{*req}
 }
 
 // HomeHandler shows popular content across the network
-func (h *FeedHandlers) HomeHandler(w http.ResponseWriter, r *http.Request) {
+func (h *BrowseHandlers) HomeHandler(w http.ResponseWriter, r *http.Request) {
 	switch r.Method {
 	case "OPTIONS":
 		util.EmptyOkHandler(w, r)
@@ -32,7 +32,7 @@ func (h *FeedHandlers) HomeHandler(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-func (h *FeedHandlers) homeHandler(w http.ResponseWriter, r *http.Request) {
+func (h *BrowseHandlers) homeHandler(w http.ResponseWriter, r *http.Request) {
 	res := map[string][]*dataset.Dataset{}
 	p := false
 	if err := h.Home(&p, &res); err != nil {
