+++ lib/feed.go
@@ -7,15 +7,15 @@ import (
 	"github.com/qri-io/dataset"
 )
 
-// FeedMethods enapsulates functions for accessing feeds of content from the
-// network. For example: a list of popular datasets is a feed.
-type FeedMethods struct {
+// BrowseMethods enapsulates functions for surveying available datasets on the
+// network.
+type BrowseMethods struct {
 	inst *Instance
 }
 
-// NewFeedMethods creates a feed instance
-func NewFeedMethods(inst *Instance) *FeedMethods {
-	m := &FeedMethods{
+// NewBrowseMethods creates a feed instance
+func NewBrowseMethods(inst *Instance) *BrowseMethods {
+	m := &BrowseMethods{
 		inst: inst,
 	}
 
@@ -23,15 +23,15 @@ func NewFeedMethods(inst *Instance) *FeedMethods {
 }
 
 // CoreRequestsName specifies this is a Methods object
-func (m *FeedMethods) CoreRequestsName() string {
-	return "feed"
+func (m *BrowseMethods) CoreRequestsName() string {
+	return "browse"
 }
 
 // Home returns a listing of datasets from a number of feeds like featured and
 // popular. Each feed is keyed by string in the response
-func (m *FeedMethods) Home(p *bool, res *map[string][]*dataset.Dataset) error {
+func (m *BrowseMethods) Home(p *bool, res *map[string][]*dataset.Dataset) error {
 	if m.inst.rpc != nil {
-		return m.inst.rpc.Call("FeedMethods.Home", p, res)
+		return m.inst.rpc.Call("BrowseMethods.Home", p, res)
 	}
 	ctx := context.TODO()
 
@@ -49,13 +49,13 @@ func (m *FeedMethods) Home(p *bool, res *map[string][]*dataset.Dataset) error {
 }
 
 // Featured asks a registry for a curated list of datasets
-func (m *FeedMethods) Featured(p *ListParams, res *[]*dataset.Dataset) error {
+func (m *BrowseMethods) Featured(p *ListParams, res *[]*dataset.Dataset) error {
 	return fmt.Errorf("featured dataset feed is not yet implemented")
 }
 
 // Recent is a feed of network datasets in reverse chronological order
 // it currently can only come from a registry, but could easily be assembled
 // via p2p methods
-func (m *FeedMethods) Recent(p *ListParams, res *[]*dataset.Dataset) error {
+func (m *BrowseMethods) Recent(p *ListParams, res *[]*dataset.Dataset) error {
 	return fmt.Errorf("recent dataset feed is not yet implemented")
 }
