+++ registry/regserver/mock_test.go
@@ -1,4 +1,4 @@
-package mock
+package regserver
 
 import (
 	"testing"
